Run project commands in an ephemeral interpreter, not the shared one
QualityMCP ran `pip install .` with sys.executable, so a target project's
dependencies landed in the interpreter ASET itself runs on. They stayed there for
later runs, several runs could mutate the same environment concurrently, and a
project declaring an incompatible pin could degrade the packages ASET needs to
work. Nine invocations shared that interpreter: pip, pytest, compileall, ruff and
pip check among them.

Each QualityMCP now builds its own ephemeral environment, outside the project
directory — apply_run passes the user's real project, not an isolated copy, so
the environment must not appear in its listings or be walked by compileall and
ruff. It inherits the toolchain through system site packages while installs
write only into itself, which costs 0.01s against 2.6s for bootstrapping pip.
A project that declares no dependencies installs nothing, so it needs no
environment at all.

Independent review rejected the first two attempts, and the findings shaped what
is here: subprocesses run with `-I` and an allowlisted environment so local
modules cannot shadow pip, pytest or ruff and inherited variables cannot redirect
installs or leak credentials; one monotonic deadline spans every phase rather
than restarting per command; a single mutation lock serialises concurrent
installs; process groups are terminated SIGTERM-then-SIGKILL, with a tree kill on
Windows; and output is bounded while draining instead of accumulating.

Four failures surfaced only from a full-suite run, which the previous session
could not perform — its host interpreter lacked chromadb, and it correctly
declined to install anything globally to get one.

A descendant that escapes the process group keeps the pipe open, so the reader
blocks in `read` awaiting EOF and closing the buffered object waits on the lock
that reader holds — the very wait the deadline exists to prevent. Releasing the
descriptor instead lets the blocked read fail immediately. The suite went from
389s to 136s.

The client's `close` raised when a worker would not stop. It runs inside `finally`
and `__exit__`, where raising replaces the caller's real exception and still
stops nothing; it now records the abandoned worker instead. That had masked a
second regression: isolating the server's cwd from the project — correct on its
own — removed the accident by which a missing workspace made Popen fail and
surface UNAVAILABLE. The condition is now checked rather than inferred.

`typing.Self` is left unused deliberately. Ruff's PYI019 autofix would introduce
it, and the project targets Python 3.10 where it does not exist — a test pins
this.

367 tests pass. Independent re-review pending; two residuals stand: resolution is
not reproducible for a project without a lockfile, and there is no network or
filesystem sandbox around the installs.

+++ src/engineering_team/mcp/server.py
@@ -75,7 +75,7 @@ def build_repository_server(root: str | Path) -> MCPServer:
     return server
 
 
-def build_quality_server(root: str | Path, timeout_seconds: int = 60) -> MCPServer:
+def build_quality_server(root: str | Path, timeout_seconds: float = 60) -> MCPServer:
     backend = QualityMCP(root, timeout_seconds=timeout_seconds)
     server = MCPServer(name="engineering-team-quality", log_level="ERROR")
 
@@ -118,7 +118,7 @@ def main() -> None:
     parser = argparse.ArgumentParser()
     parser.add_argument("--kind", choices=("repository", "quality"), required=True)
     parser.add_argument("--root", required=True)
-    parser.add_argument("--timeout", type=int, default=60)
+    parser.add_argument("--timeout", type=float, default=60)
     args = parser.parse_args()
     server = (
         build_repository_server(args.root)
